build/meson: increase timeouts for some tests
The defaults for test timeouts ias 30 seconds. That is not long
enough when running

  $ NMTST_USE_VALGRIND=1 ninja -C build test

Note that meson supports --timeout-multiplier, and automatically
increases the timeout when running under valgrind. However, that does
not work with enabling valgrind via NMTST_USE_VALGRIND=1 environment
variable.

Timeouts are really not expected for our tests. So, we can set the
timeout to a large value, because they usually are not hit.
And in the unexpected case when a test hangs, the extra wait time
does not hurt too much.