libnm: rename and expose nm_utils_base64secret_decode() in libnm
A NetworkManager client simply needs an API to validate and decode
a base64 secret -- like used for WireGuard. If we don't have this
as part of the API, it's just inconvenient. Expose it.

Rename it from _nm_utils_wireguard_decode_key(), to give it a more
general name.

Also, rename _nm_utils_wireguard_normalize_key() to
nm_utils_base64secret_normalize(). But this one we keep as
internal API. The user will care more about validating and decoding the
base64 key. To convert the key back to base64, we don't need a public
API in libnm.

This is another ABI change since 1.16-rc1.